Extract runJest and other utility functions
